+++ AsteroidGame/VisualObjects/Planet.cs
@@ -5,18 +5,19 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AsteroidGame
+namespace AsteroidGame.VisualObjects
 {
-    class Star : VisualObject
+    class Planet : VisualObject
     {
-        Image star = Image.FromFile("Star.jpg");
-        public Star(Point Position, Point Direction, int Size) 
+        private static readonly Image planet = Image.FromFile("Planet.png");
+        public Planet(Point Position, Point Direction, int Size)
             : base(Position, Direction, new Size(Size, Size))
         {
         }
+
         public override void Draw(Graphics g)
         {
-            g.DrawImage(star, _Position.X, _Position.Y, _Size.Width, _Size.Height);
+            g.DrawImage(planet, _Position.X, _Position.Y, _Size.Width, _Size.Height);
         }
 
         public override void Update()
@@ -25,10 +26,6 @@ namespace AsteroidGame
 
             if (_Position.X < 0)
                 _Position.X = Game.Width + _Size.Width;
-
-            //if (_Position.X > Game.Width /*+ _Size.Width*/)
-            //    _Position.X = 0 - _Size;
         }
     }
 }
-
