+++ AsteroidGame/VisualObjects/VisualObject.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using System.Drawing;
 using System.Windows.Forms;
 
-namespace AsteroidGame
+namespace AsteroidGame.VisualObjects
 {
     abstract class VisualObject
     {
